LoggerFactory and loggers now use properties for their initial log level
Constructor Injected Services don't have the option to use the implementation
before it is started as their constructor gets called when they're starting.
LoggerFactory therefore had to be refactored to use properties instead of
available methods on interfaces.

Having these methods exposed on regular services is a hazard as well,
but it provides a really good ergonomy for e.g. timers and so far
does not show to be a problem. Time will tell?