Add coroutine type Task
Task is a lighter-weight and less boilerplate-y way for users to define
and use coroutine functions. Added benefit is that it removes the need
to include <AsyncGenerator.h>, which pulls in a lot of other headers.

Also remove the use of <atomic> in all coroutine classes, as Ichor takes
care to confine threads, making its use unnecessary.